requirements.txt
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.3.0
openai>=1.0.0
langgraph>=0.2.0
langgraph-supervisor>=0.0.1
chromadb>=0.5.0
gradio>=5.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
mcp>=1.0.0
pytest>=8.0.0
python-multipart>=0.0.9
python-telegram-bot>=21.0
opencv-python>=4.8.0
streamlit>=1.38.0
